backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
google-generativeai==0.3.0
openai==1.3.8
python-telegram-bot==20.3
requests==2.31.0
cryptography==41.0.7
PyJWT==2.8.1
python-multipart==0.0.6
